wac: connect to Home Assistent via WebSocket
This is rather complicated, as we need to use incremental IDs in the
WebSocket messages, so when the initial attempt fails, we need to use a
new ID but keep track of the initial id, etc.

This code is horrible, and doesn't measure timing yet, but it does
appear to work. We're also not cleaning up anything from the session
tracker yet.